requirements.txt
asgiref
certifi
cffi
charset-normalizer
cryptography
Django
django-allauth
django-cors-headers
djangorestframework
djangorestframework_simplejwt
idna
pillow
pycparser
PyJWT
requests
sqlparse
tzdata
urllib3